+++ lib/core/service_locator.dart
@@ -1,5 +1,6 @@
 import 'package:get_it/get_it.dart';
 import 'package:stockflow/transactions/bloc/balance_cubit.dart';
+import 'package:stockflow/transactions/bloc/categories_bloc.dart';
 import 'package:stockflow/transactions/bloc/category_totals_cubit.dart';
 import 'package:stockflow/transactions/bloc/transactions_bloc.dart';
 import 'package:stockflow/transactions/data/connection.dart';
@@ -8,11 +9,14 @@ import 'package:stockflow/transactions/data/repos/transactions_repository_impl.d
 import 'package:stockflow/transactions/data/transactions_data_source.dart';
 import 'package:stockflow/transactions/domain/repositories/category_repository.dart';
 import 'package:stockflow/transactions/domain/repositories/transactions_repository.dart';
+import 'package:stockflow/transactions/domain/usecases/add_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/add_transaction_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/delete_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/delete_transaction_usecase.dart';
-import 'package:stockflow/transactions/domain/usecases/get_categories_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/get_transactions_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/update_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/watch_balance_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/watch_categories_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/watch_category_totals_usecase.dart';
 
 final getIt = GetIt.instance;
@@ -35,8 +39,20 @@ void setupServiceLocator() {
     () => GetTransactionsUseCase(getIt<TransactionsRepository>()),
   );
 
-  getIt.registerLazySingleton<GetCategoriesUseCase>(
-    () => GetCategoriesUseCase(getIt<CategoryRepository>()),
+  getIt.registerLazySingleton<WatchCategoriesUseCase>(
+    () => WatchCategoriesUseCase(getIt<CategoryRepository>()),
+  );
+
+  getIt.registerLazySingleton<AddCategoryUseCase>(
+    () => AddCategoryUseCase(getIt<CategoryRepository>()),
+  );
+
+  getIt.registerLazySingleton<UpdateCategoryUseCase>(
+    () => UpdateCategoryUseCase(getIt<CategoryRepository>()),
+  );
+
+  getIt.registerLazySingleton<DeleteCategoryUseCase>(
+    () => DeleteCategoryUseCase(getIt<CategoryRepository>()),
   );
 
   getIt.registerLazySingleton<WatchBalanceUseCase>(
@@ -62,7 +78,7 @@ void setupServiceLocator() {
       getTransactionsUseCase: getIt<GetTransactionsUseCase>(),
       addTransactionUseCase: getIt<AddTransactionUseCase>(),
       deleteTransactionUseCase: getIt<DeleteTransactionUseCase>(),
-      getCategoriesUseCase: getIt<GetCategoriesUseCase>(),
+      watchCategoriesUseCase: getIt<WatchCategoriesUseCase>(),
     ),
   );
 
@@ -73,6 +89,15 @@ void setupServiceLocator() {
   getIt.registerFactory<CategoryTotalsCubit>(
     () => CategoryTotalsCubit(
       watchCategoryTotals: getIt<WatchCategoryTotalsUsecase>(),
+    ),
+  );
+
+  getIt.registerFactory<CategoriesBloc>(
+    () => CategoriesBloc(
+      watchCategoriesUseCase: getIt<WatchCategoriesUseCase>(),
+      addCategoryUseCase: getIt<AddCategoryUseCase>(),
+      updateCategoryUseCase: getIt<UpdateCategoryUseCase>(),
+      deleteCategoryUseCase: getIt<DeleteCategoryUseCase>(),
     ),
   );
 }