+++ lib/transactions/data/models/transactions.dart
@@ -9,7 +9,11 @@ class Transactions extends Table {
   TextColumn get note => text().nullable()();
   late final occurredTime = dateTime().withDefault(currentDateAndTime)();
   late final creationTime = dateTime().withDefault(currentDateAndTime)();
-  TextColumn get categoryId => text().nullable().references(Categories, #id)();
+  TextColumn get categoryId => text().nullable().references(
+    Categories,
+    #id,
+    onDelete: KeyAction.setNull,
+  )();
 
   @override
   Set<Column> get primaryKey => {id};