+++ test/helpers/mocks.dart
@@ -1,11 +1,14 @@
 import 'package:mocktail/mocktail.dart';
 import 'package:stockflow/transactions/domain/repositories/category_repository.dart';
 import 'package:stockflow/transactions/domain/repositories/transactions_repository.dart';
+import 'package:stockflow/transactions/domain/usecases/add_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/add_transaction_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/delete_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/delete_transaction_usecase.dart';
-import 'package:stockflow/transactions/domain/usecases/get_categories_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/get_transactions_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/update_category_usecase.dart';
 import 'package:stockflow/transactions/domain/usecases/watch_balance_usecase.dart';
+import 'package:stockflow/transactions/domain/usecases/watch_categories_usecase.dart';
 
 class MockTransactionsRepository extends Mock
     implements TransactionsRepository {}
@@ -20,6 +23,13 @@ class MockAddTransactionUseCase extends Mock implements AddTransactionUseCase {}
 class MockDeleteTransactionUseCase extends Mock
     implements DeleteTransactionUseCase {}
 
-class MockGetCategoriesUseCase extends Mock implements GetCategoriesUseCase {}
+class MockWatchCategoriesUseCase extends Mock
+    implements WatchCategoriesUseCase {}
+
+class MockAddCategoryUseCase extends Mock implements AddCategoryUseCase {}
+
+class MockUpdateCategoryUseCase extends Mock implements UpdateCategoryUseCase {}
+
+class MockDeleteCategoryUseCase extends Mock implements DeleteCategoryUseCase {}
 
 class MockWatchBalanceUseCase extends Mock implements WatchBalanceUseCase {}